+++ apps/vs-code-designer-react/src/app/DesignerCommandBar/index.tsx
@@ -16,7 +16,7 @@ export const DesignerCommandBar: React.FC<DesignerCommandBarProps> = ({ isMonito
 
   const onSave = async () => {
     const designerState = DesignerStore.getState();
-    const { definition, parameters } = await serializeBJSWorkflow(designerState, {
+    const { definition, parameters, connectionReferences } = await serializeBJSWorkflow(designerState, {
       skipValidation: true,
       ignoreNonCriticalErrors: true,
     });
@@ -24,6 +24,7 @@ export const DesignerCommandBar: React.FC<DesignerCommandBarProps> = ({ isMonito
       command: ExtensionCommand.save,
       definition,
       parameters,
+      connectionReferences,
     });
   };