+++ apps/vs-code-designer-react/src/app/app.tsx
@@ -1,19 +1,37 @@
-import type { RootState } from '../state/Store';
+import { createFileSystemConnection } from '../state/DesignerSlice';
+import type { AppDispatch, RootState } from '../state/Store';
 import { VSCodeContext } from '../webviewCommunication';
 import { DesignerCommandBar } from './DesignerCommandBar';
 import { getDesignerServices } from './servicesHelper';
+import { convertConnectionsDataToReferences } from './utilities/workflow';
+import type { ConnectionCreationInfo } from '@microsoft/designer-client-services-logic-apps';
+import type { ConnectionReferences } from '@microsoft/logic-apps-designer';
 import { DesignerProvider, BJSWorkflowProvider, Designer, getTheme, useThemeObserver } from '@microsoft/logic-apps-designer';
 import { Theme } from '@microsoft/utils-logic-apps';
-import { useContext, useEffect, useMemo, useState } from 'react';
-import { useSelector } from 'react-redux';
+import type { FileSystemConnectionInfo } from '@microsoft/vscode-extension';
+import { ExtensionCommand } from '@microsoft/vscode-extension';
+import { useContext, useMemo, useState, useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
 
 export const App = () => {
-  const vscodeState = useSelector((state: RootState) => state.designer);
-  const { panelMetaData, connectionReferences, baseUrl, apiHubServiceDetails, readOnly, isLocal, apiVersion, isMonitoringView, runId } =
-    vscodeState;
-  const codelessApp = panelMetaData?.codelessApp;
-  const [theme, setTheme] = useState<Theme>(getTheme(document.body));
   const vscode = useContext(VSCodeContext);
+  const dispatch: AppDispatch = useDispatch();
+  const vscodeState = useSelector((state: RootState) => state.designer);
+  const {
+    panelMetaData,
+    connectionData,
+    baseUrl,
+    apiHubServiceDetails,
+    readOnly,
+    isLocal,
+    apiVersion,
+    tenantId,
+    oauthRedirectUrl,
+    isMonitoringView,
+    runId,
+  } = vscodeState;
+  const standardApp = panelMetaData?.standardApp;
+  const [theme, setTheme] = useState<Theme>(getTheme(document.body));
   const [runInstance, setRunInstance] = useState({});
 
   useThemeObserver(document.body, theme, setTheme, {
@@ -21,8 +39,36 @@ export const App = () => {
   });
 
   const services = useMemo(() => {
-    return getDesignerServices(baseUrl, apiVersion, apiHubServiceDetails, isLocal, vscode);
-  }, [baseUrl, apiVersion, apiHubServiceDetails, isLocal]);
+    const fileSystemConnectionCreate = async (
+      connectionInfo: FileSystemConnectionInfo,
+      connectionName: string
+    ): Promise<ConnectionCreationInfo> => {
+      vscode.postMessage({
+        command: ExtensionCommand.createFileSystemConnection,
+        connectionInfo,
+        connectionName,
+      });
+      return new Promise((resolve, reject) => {
+        dispatch(createFileSystemConnection({ connectionName, resolve, reject }));
+      });
+    };
+    return getDesignerServices(
+      baseUrl,
+      apiVersion,
+      apiHubServiceDetails,
+      tenantId,
+      isLocal,
+      connectionData,
+      panelMetaData,
+      fileSystemConnectionCreate,
+      vscode,
+      oauthRedirectUrl
+    );
+  }, [baseUrl, apiVersion, apiHubServiceDetails, tenantId, isLocal, connectionData, panelMetaData, vscode, oauthRedirectUrl, dispatch]);
+
+  const connectionReferences: ConnectionReferences = useMemo(() => {
+    return convertConnectionsDataToReferences(connectionData);
+  }, [connectionData]);
 
   useEffect(() => {
     async function getRunInstance() {
@@ -45,8 +91,8 @@ export const App = () => {
         services: services,
       }}
     >
-      {codelessApp ? (
-        <BJSWorkflowProvider workflow={{ definition: codelessApp.definition, connectionReferences }}>
+      {standardApp ? (
+        <BJSWorkflowProvider workflow={{ definition: standardApp.definition, connectionReferences }}>
           {readOnly ? null : <DesignerCommandBar isMonitoringView />}
           <Designer />
         </BJSWorkflowProvider>