+++ apps/vs-code-designer-react/src/app/servicesHelper.ts
@@ -1,106 +1,110 @@
-import { HttpClient } from './httpClient';
+import { clientSupportedOperations } from './constants';
+import { HttpClient } from './services/httpClient';
+import { StandardOAuthService } from './services/oAuth';
+import { resolveConnectionsReferences } from './utilities/workflow';
 import {
   StandardConnectionService,
   StandardConnectorService,
   StandardOperationManifestService,
   StandardSearchService,
-  StandardOAuthService,
   StandardGatewayService,
   StandardRunService,
 } from '@microsoft/designer-client-services-logic-apps';
-import type { IApiHubServiceDetails, ContentType, IHostService, IWorkflowService } from '@microsoft/designer-client-services-logic-apps';
-import { ResourceIdentityType, HTTP_METHODS } from '@microsoft/utils-logic-apps';
+import type { IApiHubServiceDetails, ConnectionCreationInfo, ContentType } from '@microsoft/designer-client-services-logic-apps';
+import { HTTP_METHODS } from '@microsoft/utils-logic-apps';
+import type { ConnectionAndAppSetting, ConnectionsData, IDesignerPanelMetadata } from '@microsoft/vscode-extension';
 import { ExtensionCommand } from '@microsoft/vscode-extension';
 import type { WebviewApi } from 'vscode-webview';
-
-const httpClient = new HttpClient();
 
 export const getDesignerServices = (
   baseUrl: string,
   apiVersion: string,
   apiHubServiceDetails: IApiHubServiceDetails,
+  tenantId: string | undefined,
   isLocal: boolean,
-  vscode: WebviewApi<unknown>
-): {
-  connectionService: StandardConnectionService;
-  connectorService: StandardConnectorService;
-  operationManifestService: StandardOperationManifestService;
-  searchService: StandardSearchService;
-  oAuthService: StandardOAuthService;
-  gatewayService: StandardGatewayService;
-  workflowService: IWorkflowService;
-  hostService: IHostService;
-  runService: StandardRunService;
-} => {
+  connectionData: ConnectionsData,
+  panelMetadata: IDesignerPanelMetadata | null,
+  createFileSystemConnection: (connectionInfo: ConnectionCreationInfo, connectionName: string) => Promise<ConnectionCreationInfo>,
+  vscode: WebviewApi<unknown>,
+  oauthRedirectUrl: string
+): any => {
+  let authToken = '',
+    panelId = '',
+    workflowDetails: Record<string, any> = {},
+    appSettings = {};
+
+  if (panelMetadata) {
+    authToken = panelMetadata.accessToken ?? '';
+    panelId = panelMetadata.panelId;
+    workflowDetails = panelMetadata.workflowDetails;
+    appSettings = panelMetadata.localSettings;
+  }
+
+  const addConnectionData = async (connectionAndSetting: ConnectionAndAppSetting): Promise<void> => {
+    return vscode.postMessage({
+      command: ExtensionCommand.addConnection,
+      connectionAndSetting,
+    });
+  };
+
+  const httpClient = new HttpClient({ accessToken: authToken, baseUrl, apiHubBaseUrl: apiHubServiceDetails.baseUrl });
   const connectionService = new StandardConnectionService({
     baseUrl,
     apiVersion,
     httpClient,
     apiHubServiceDetails,
-    workflowAppDetails: { appName: 'app', identity: { type: ResourceIdentityType.SYSTEM_ASSIGNED } },
-    readConnections: () => Promise.resolve({}),
+    tenantId,
+    readConnections: () => Promise.resolve(connectionData),
+    writeConnection: (connectionAndSetting: ConnectionAndAppSetting) => {
+      return addConnectionData(connectionAndSetting);
+    },
+    connectionCreationClients: {
+      FileSystem: {
+        connectionCreationFunc: createFileSystemConnection,
+      },
+    },
   });
+
+  const manualWorkflows = Object.keys(workflowDetails).map((name) => ({ value: name, displayName: name }));
 
   const connectorService = new StandardConnectorService({
     apiVersion,
     baseUrl,
     httpClient,
-    apiHubServiceDetails,
-    clientSupportedOperations: [
-      {
-        connectorId: 'connectionProviders/localWorkflowOperation',
-        operationId: 'invokeWorkflow',
-      },
-      {
-        connectorId: 'connectionProviders/xmlOperations',
-        operationId: 'xmlValidation',
-      },
-      {
-        connectorId: 'connectionProviders/xmlOperations',
-        operationId: 'xmlTransform',
-      },
-      {
-        connectorId: 'connectionProviders/liquidOperations',
-        operationId: 'liquidJsonToJson',
-      },
-      {
-        connectorId: 'connectionProviders/liquidOperations',
-        operationId: 'liquidJsonToText',
-      },
-      {
-        connectorId: 'connectionProviders/liquidOperations',
-        operationId: 'liquidXmlToJson',
-      },
-      {
-        connectorId: 'connectionProviders/liquidOperations',
-        operationId: 'liquidXmlToText',
-      },
-      {
-        connectorId: 'connectionProviders/flatFileOperations',
-        operationId: 'flatFileDecoding',
-      },
-      {
-        connectorId: 'connectionProviders/flatFileOperations',
-        operationId: 'flatFileEncoding',
-      },
-      {
-        connectorId: 'connectionProviders/swiftOperations',
-        operationId: 'SwiftDecode',
-      },
-      {
-        connectorId: 'connectionProviders/swiftOperations',
-        operationId: 'SwiftEncode',
-      },
-    ],
-    getConfiguration: () => Promise.resolve({}),
-    schemaClient: {
-      getWorkflowSwagger: () => Promise.resolve({}),
+    clientSupportedOperations: clientSupportedOperations,
+    getConfiguration: async (connectionId: string): Promise<any> => {
+      if (!connectionId) {
+        return Promise.resolve();
+      }
+
+      const connectionName = connectionId.split('/').splice(-1)[0];
+      const connectionInfo = connectionData?.serviceProviderConnections?.[connectionName];
+
+      if (connectionInfo) {
+        const resolvedConnectionInfo = resolveConnectionsReferences(JSON.stringify(connectionInfo), {}, appSettings);
+        delete resolvedConnectionInfo.displayName;
+
+        return {
+          connection: resolvedConnectionInfo,
+        };
+      }
+
+      return undefined;
     },
+    schemaClient: {
+      getWorkflowSwagger: (args) => {
+        const workflowName = args.parameters.name;
+        const workflowSchemas = JSON.stringify(workflowDetails);
+        return Promise.resolve(workflowSchemas[workflowName] || {});
+      },
+    },
+    // TODO ValuesClient needs to be implemented
     valuesClient: {
-      getWorkflows: () => Promise.resolve([]),
+      getWorkflows: () => Promise.resolve(manualWorkflows),
       getMapArtifacts: () => Promise.resolve([]),
       getSchemaArtifacts: () => Promise.resolve([]),
     },
+    apiHubServiceDetails,
     workflowReferenceId: '',
   });
 
@@ -115,27 +119,34 @@ export const getDesignerServices = (
     apiVersion,
     httpClient,
     apiHubServiceDetails,
-    isDev: true,
+    isDev: false,
   });
 
+  const { subscriptionId, resourceGroup, location } = apiHubServiceDetails;
+
   const oAuthService = new StandardOAuthService({
-    baseUrl,
-    apiVersion,
+    vscode,
+    panelId,
+    authToken,
+    oauthRedirectUrl,
+    baseUrl: apiHubServiceDetails.baseUrl,
+    apiVersion: '2018-07-01-preview',
     httpClient,
-    subscriptionId: '',
-    resourceGroup: '',
-    location: '',
+    subscriptionId,
+    resourceGroup,
+    location,
   });
 
   const gatewayService = new StandardGatewayService({
     baseUrl,
     httpClient,
     apiVersions: {
-      subscription: '2018-11-01',
+      subscription: apiVersion,
       gateway: '2016-06-01',
     },
   });
 
+  // Workflow service needs to be implemented to get the callback url for azure resources
   const workflowService = {
     getCallbackUrl: async () => {
       if (isLocal) {