+++ libs/designer/src/lib/core/utils/parameters/helper.ts
@@ -2728,8 +2728,6 @@ export function validateUntilAction(
   parameters: ParameterInfo[],
   changedParameter: Partial<ParameterInfo>
 ) {
-  console.log('### Changed', changedParameter);
-
   const errorMessage = 'Either limit count or timout must be specified.';
 
   const countParameter = parameters.find((parameter) => parameter.parameterName === 'limit.count');