+++ libs/vscode-extension/src/lib/models/extensioncommand.ts
@@ -12,8 +12,13 @@ export enum ExtensionCommand {
   save = 'Save',
   showContent = 'ShowContent',
   resubmitRun = 'ResubmitRun',
+  addConnection = 'add-connection',
+  createFileSystemConnection = 'create-file-system-connection',
+  completeFileSystemConnection = 'complete-file-system-connection',
   getCallbackUrl = 'GetCallbackUrl',
   receiveCallback = 'ReceiveCallback',
+  openOauthLoginPopup = 'OpenLoginPopup',
+  completeOauthLogin = 'CompleteOauthLogin',
 }
 
 export interface IExtensionsJson {